User Points Recycler View Added

+++ app/src/main/java/com/guptas/android/model/Users.java
@@ -3,16 +3,17 @@ package com.guptas.android.model;
 /**
  * Created by sakshiagarwal on 09/04/16.
  */
+
 public class Users {
     private String name;
-    private Integer points;
+    private Long points;
 
     // Required default constructor for Firebase object mapping
     @SuppressWarnings("unused")
     public Users() {
     }
 
-    public Users(String name, Integer points)
+    public Users(String name, Long points)
     {
         this.name = name;
         this.points = points;
@@ -22,7 +23,7 @@ public class Users {
         return name;
     }
 
-    public Integer getPoints() {
+    public Long getPoints() {
         return points;
     }
 }
